Move MockThreadScheduler to test utils and add instance to ConfigurationLoaderUnitTest.
