+++ SharedUtils/src/test/java/com/braintreepayments/api/sharedutils/HttpClientUnitTest.java
@@ -5,10 +5,11 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
+
+import com.braintreepayments.api.testutils.MockThreadScheduler;
 
 import org.junit.Before;
 import org.junit.Test;
